pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tax-agent"
version = "0.1.0"
description = "개인소득세 AI 세무 에이전트 (CLI)"
readme = "CLAUDE.md"
requires-python = ">=3.12"
dependencies = [
  "pdfplumber>=0.11.0",
  "python-dotenv>=1.0.0",
  "httpx>=0.27.0",
  "rich>=13.7.0",
  "fpdf2>=2.8.7",
  "pyyaml>=6.0.3",
]

[project.scripts]
tax-agent = "main:main"
tax-agent-smoke = "qa:smoke"
tax-agent-law-watch = "law_watch:main"

[tool.setuptools]
# 변경 이유:
# - 현재 레포는 패키지 디렉터리(src/패키지명) 구조가 아닌 단일 모듈 파일 구조다.
# - [project.scripts] 엔트리포인트가 동작하려면 실행 대상 모듈이 배포 메타데이터에 포함되어야 한다.
py-modules = [
  "main",
  "qa",
  "tax_calculator",
  "tax_store",
  "strategy_engine_legacy",
  "document_parser",
  "legal_search",
  "law_watch",
]

[tool.setuptools.packages.find]
include = ["strategy_engine*"]
